Reverted change for 5ghz minbitrate back to int? from double?

+++ Meraki.Api/Data/FiveGhzSettings.cs
@@ -25,7 +25,7 @@ public class FiveGhzSettings
 	/// </summary>
 	[ApiAccess(ApiAccess.ReadWrite)]
 	[DataMember(Name = "minBitrate")]
-	public double? MinBitrate { get; set; }
+	public int? MinBitrate { get; set; }
 
 	/// <summary>
 	/// Sets valid auto channels for 5Ghz band. Can be one of '36', '40', '44', '48', '52', '56', '60', '64', '100', '104', '108', '112', '116', '120', '124', '128', '132', '136', '140', '144', '149', '153', '157', '161' or '165'.Defaults to [36, 40, 44, 48, 52, 56, 60, 64, 100, 104, 108, 112, 116, 120, 124, 128, 132, 136, 140, 144, 149, 153, 157, 161, 165].
